+++ src/dotnet/Phosphor/PhosphorTypewriter.cs
@@ -1,7 +1,5 @@
 ﻿using System.Threading.Channels;
-using Enclave.Common;
 using Enclave.Common.Helpers;
-using Enclave.Shared.IO;
 
 namespace Enclave.Phosphor;